device: update i2c class with working revision
The i2c class is at the moment only composed of a transfer structure that
represents a read of write transfer of data between the master and the slave.
The master is here the entity that runs MutekH and the slave is an external
device.

The transfer specifies three major information: 1) slave address, 2) slave
register address to be accessed and 3) the data buffer to read into of write
from.

The i2c class provide also a low-level blocking api for scanning, reading and
writing. The scan method returns if the given slave address matches with a
slave connected on the bus.